Move SpaceViewBlueprint -> re_space_view
