+++ crates/re_space_view/Cargo.toml
@@ -19,10 +19,14 @@ all-features = true
 default = []
 
 [dependencies]
-re_entity_db.workspace = true
+re_log.workspace = true
 re_log_types.workspace = true
+re_data_store.workspace = true
+re_entity_db.workspace = true
+re_query.workspace = true
 re_tracing.workspace = true
 re_types_core.workspace = true
+re_types.workspace = true
 re_viewer_context.workspace = true
 
 ahash.workspace = true
