+++ crates/re_space_view/src/lib.rs
@@ -8,11 +8,13 @@ mod data_query;
 mod data_query_blueprint;
 mod heuristics;
 mod screenshot;
+mod space_view;
 
 pub use data_query::{DataQuery, EntityOverrideContext, PropertyResolver};
 pub use data_query_blueprint::DataQueryBlueprint;
 pub use heuristics::suggest_space_view_for_each_entity;
 pub use screenshot::ScreenshotMode;
+pub use space_view::{SpaceViewBlueprint, SpaceViewName};
 
 // -----------
