+++ crates/re_viewport/src/add_space_view_or_container_modal.rs
@@ -2,9 +2,9 @@
 
 use itertools::Itertools;
 
-use crate::{icon_for_container_kind, SpaceViewBlueprint, ViewportBlueprint};
+use crate::{icon_for_container_kind, ViewportBlueprint};
 use re_log_types::{EntityPath, EntityPathFilter};
-use re_space_view::DataQueryBlueprint;
+use re_space_view::{DataQueryBlueprint, SpaceViewBlueprint};
 use re_ui::ReUi;
 use re_viewer_context::{ContainerId, ViewerContext};
