+++ crates/re_viewport/src/auto_layout.rs
@@ -8,7 +8,7 @@ use itertools::Itertools as _;
 
 use re_viewer_context::{SpaceViewClassIdentifier, SpaceViewId};
 
-use super::space_view::SpaceViewBlueprint;
+use re_space_view::SpaceViewBlueprint;
 
 #[derive(Clone, Debug)]
 struct SpaceMakeInfo {
