+++ crates/re_viewport/src/space_view_entity_picker.rs
@@ -4,9 +4,10 @@ use nohash_hasher::IntMap;
 use re_data_ui::item_ui;
 use re_entity_db::{EntityPath, EntityTree, InstancePath};
 use re_log_types::{EntityPathFilter, EntityPathRule};
+use re_space_view::SpaceViewBlueprint;
 use re_viewer_context::{DataQueryResult, SpaceViewId, ViewerContext};
 
-use crate::{determine_visualizable_entities, space_view::SpaceViewBlueprint, ViewportBlueprint};
+use crate::{determine_visualizable_entities, ViewportBlueprint};
 
 /// Window for adding/removing entities from a space view.
 ///
