+++ crates/re_viewport/src/space_view_heuristics.rs
@@ -1,7 +1,7 @@
 use re_space_view::DataQueryBlueprint;
 use re_viewer_context::ViewerContext;
 
-use crate::space_view::SpaceViewBlueprint;
+use re_space_view::SpaceViewBlueprint;
 
 /// List out all space views we generate by default for the available data.
 ///
